# grops-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface GraphSourceMeta {
  id: string;
  name: string;
  pages: number;
  charCount: number;
}

export interface GraphNode {
  id: string;
  label: string;
  topic: string;
  color: string;
  summary: string;
  quote: string;
  sourceDocId?: string | null;
  /** start | concept | finish */
  kind: string;
}

export interface GraphEdge {
  from: string;
  to: string;
  label: string;
  color: string;
  /** cause | correlation */
  kind: string;
}

export interface TopicLegendEntry {
  topic: string;
  color: string;
  count: number;
}

export interface Neurograph {
  id: string;
  title: string;
  /** summary | detailed */
  mode: string;
  sources: GraphSourceMeta[];
  createdAt: string;
  nodes: GraphNode[];
  edges: GraphEdge[];
  topics: TopicLegendEntry[];
  generationMs: number;
}

export interface GraphSummary {
  id: string;
  title: string;
  mode: string;
  createdAt: string;
  nodeCount: number;
  edgeCount: number;
  sourceCount: number;
}

export interface UsageStatus {
  used: number;
  limit: number;
}

export interface ClarifyBody {
  /** @minLength 1 */
  question: string;
}

export interface ClarifyResult {
  answer: string;
  highlightedNodeIds: string[];
}

export interface ExpandBody {
  /** @minLength 1 */
  newInfo: string;
}

export interface ExpansionProposal {
  rationale: string;
  newNodes: GraphNode[];
  newEdges: GraphEdge[];
}

export interface WaitlistBody {
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface WaitlistStatus {
  /** added | already_on_waitlist */
  status: string;
}

export interface AdminUserRow {
  userId: string;
  graphsThisMonth: number;
  totalGraphs: number;
  lastActivity: string;
  limit: number;
}

export interface AdminUsersResponse {
  monthStart: string;
  totalUsers: number;
  users: AdminUserRow[];
}

export interface AdminWaitlistEntry {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface AdminWaitlistResponse {
  total: number;
  entries: AdminWaitlistEntry[];
}

export interface AdminHealthResponse {
  total: number;
  successes: number;
  failures: number;
  successRate: number;
  avgMs: number;
}