# grops-main/lib/api-zod/src/generated/types/adminUsersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */
import type { AdminUserRow } from "./adminUserRow";

export interface AdminUsersResponse {
  monthStart: Date;
  totalUsers: number;
  users: AdminUserRow[];
}