# grops-main/lib/api-zod/src/generated/types/expansionProposal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */
import type { GraphEdge } from "./graphEdge";
import type { GraphNode } from "./graphNode";

export interface ExpansionProposal {
  rationale: string;
  newNodes: GraphNode[];
  newEdges: GraphEdge[];
}