# grops-main/lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */

export interface GraphEdge {
  from: string;
  to: string;
  label: string;
  color: string;
  /** cause | correlation */
  kind: string;
}