# grops-main/lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */

export interface GraphNode {
  id: string;
  label: string;
  topic: string;
  color: string;
  summary: string;
  quote: string;
  sourceDocId?: string | null;
  /** start | concept | finish */
  kind: string;
}