# grops-main/lib/api-zod/src/generated/types/graphSourceMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */

export interface GraphSourceMeta {
  id: string;
  name: string;
  pages: number;
  charCount: number;
}