# grops-main/lib/api-zod/src/generated/types/neurograph.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grops V1 API
 * OpenAPI spec version: 1.0.0
 */
import type { GraphEdge } from "./graphEdge";
import type { GraphNode } from "./graphNode";
import type { GraphSourceMeta } from "./graphSourceMeta";
import type { TopicLegendEntry } from "./topicLegendEntry";

export interface Neurograph {
  id: string;
  title: string;
  /** summary | detailed */
  mode: string;
  sources: GraphSourceMeta[];
  createdAt: Date;
  nodes: GraphNode[];
  edges: GraphEdge[];
  topics: TopicLegendEntry[];
  generationMs: number;
}